+++ packages/decorators/find-replace/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-find-replace
 
+## 18.1.1
+
 ## 17.0.3
 
 ## 17.0.2