+++ packages/nodes/basic-elements/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/plate-basic-elements
 
+## 18.1.1
+
 ## 17.0.3
 
 ## 17.0.2